refactor: extract NavigationStateContext
